mysite/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-dotenv==1.4.2
django-filter==24.3
django-modeltranslation==0.19.11
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
MarkupSafe==3.0.2
msgpack==1.1.0
openapi-codec==1.3.2
packaging==24.2
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
simplejson==3.19.3
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zope.interface==7.2
